+++ perseo-quality/docs/source/changelog.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+v1.0.10
+-------
+
+**Bug Fixes**
+
+- Fixing bug in Point Target Analysis channel name output type, now properly converted to string
+
 v1.0.9
 ------